requirements.txt
pandas
numpy
matplotlib
sentence-transformers
seaborn
faiss-cpu
scikit-learn
streamlit
jupyterlab
fastapi